requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
google-generativeai==0.3.1
yfinance==0.2.33
twstock==1.3.1
fredapi==0.5.1
